+++ app/exceptions/vinyl_exception.py
@@ -1,9 +1,17 @@
-from app.exceptions.base_exception import NotFoundError, AlreadyExistsError, ForbiddenError
+from app.exceptions.base_exception import NotFoundError, AlreadyExistsError, ForbiddenError, PendingError, Registered
 
 class VinylNotFoundError(NotFoundError):
     def __init__(self, vinyl_id: int):
         super().__init__("Vinyl", vinyl_id)
 
 class VinylForbiddenError(ForbiddenError):
+    def __init__(self, vinyl_id: int, user_id: int):
+        super().__init__("Vinyl", vinyl_id, user_id)
+
+class VinylPending(PendingError):
+    def __init__(self, vinyl_id: int, user_id: int):
+        super().__init__("Vinyl", vinyl_id, user_id)
+
+class VinylCreated(Registered):
     def __init__(self, vinyl_id: int, user_id: int):
         super().__init__("Vinyl", vinyl_id, user_id)