+++ app/main.py
@@ -3,6 +3,7 @@ from app.controllers.auth_controller import router as auth_router
 from app.controllers.device_controller import router as device_router
 from app.controllers.vinyl_controller import router as vinyl_router
 from app.controllers.resource_controller import router as resource_router
+from app.controllers.play_controller import router as play_router
 
 app = FastAPI(title="Orpheus API")
 
@@ -14,3 +15,4 @@ app.include_router(auth_router)
 app.include_router(device_router)
 app.include_router(vinyl_router)
 app.include_router(resource_router)
+app.include_router(play_router)